# backend/app/routers/policies.py
"""
Policies Router
===============

Policy CRUD and management endpoints.
"""

from fastapi import APIRouter, HTTPException
from typing import List

from shared.dto.policy_dto import PolicyCreateRequestDTO, PolicyResponseDTO, PolicyListResponseDTO

router = APIRouter()


@router.get("/", response_model=PolicyListResponseDTO)
async def list_policies():
    """
    List all policies.
    
    Returns:
        PolicyListResponseDTO with all policies
        
    TODO: Implement policy listing
    """
    # TODO: Get policies from repository
    return PolicyListResponseDTO()


@router.post("/", response_model=PolicyResponseDTO)
async def create_policy(request: PolicyCreateRequestDTO):
    """
    Create a new policy.
    
    Args:
        request: Policy creation request
        
    Returns:
        Created policy
        
    TODO: Implement policy creation
    """
    # TODO: Create policy and apply to simulation
    return PolicyResponseDTO()


@router.get("/{policy_id}", response_model=PolicyResponseDTO)
async def get_policy(policy_id: str):
    """
    Get a specific policy.
    
    Args:
        policy_id: Policy ID
        
    Returns:
        PolicyResponseDTO
        
    TODO: Implement policy retrieval
    """
    # TODO: Get policy from repository
    return PolicyResponseDTO()


@router.delete("/{policy_id}")
async def revoke_policy(policy_id: str):
    """
    Revoke a policy.
    
    Args:
        policy_id: Policy ID to revoke
        
    Returns:
        Success status
        
    TODO: Implement policy revocation
    """
    # TODO: Revoke policy
    return {"status": "revoked"}